Extract Servant API types into aftok-api package
Create a new aftok-api package in api/ containing pure Servant API type
definitions extracted from aftok-executables. This separates API types
from server implementation, allowing clients to depend on API types
without pulling in server dependencies.

New package structure:
- Aftok.API: Top-level re-exports, AftokAPI, VersionedAPI, ProtectedAPI
- Aftok.API.Types: HttpApiData/JSON instances for ID types
- Aftok.API.Auth: AuthenticatedUser, LoginRequest, AftokAuth
- Aftok.API.Users: UsersAPI, registration and captcha types
- Aftok.API.Session: SessionAPI, LoginCheckResponse
- Aftok.API.Projects: ProjectsAPI, Contributor, ProjectDetail
- Aftok.API.WorkLog: WorkLogAPI, log request types
- Aftok.API.Auctions: AuctionsAPI, auction/bid request types
- Aftok.API.Billing: BillingAPI, billable request types
- Aftok.API.Payments: PaymentsAPI
- Aftok.API.PasswordReset: PasswordResetAPI, request/response types
- Aftok.API.Config: ConfigAPI, ClientConfig

Server modules updated to import from aftok-api and retain only handlers.

Benefits:
- Client generation: Clients can depend on aftok-api without server deps
- Type sharing: Consistent types between server and clients
- Smaller dependencies: API consumers don't need database, email, etc.
- Documentation: API types serve as documentation
- Testing: Easier to test API contracts
