+++ executables/aftok-executables.cabal
@@ -88,7 +88,6 @@ Executable aftok-server
     Aftok.Servant.Auth
     Aftok.Servant.API
     Aftok.Servant.Config
-    Aftok.Servant.Instances
     Aftok.Servant.Json
     Aftok.Servant.PasswordReset
     Aftok.Servant.Server
@@ -102,6 +101,7 @@ Executable aftok-server
 
   build-depends:
     aftok,
+    aftok-api,
     directory                 >= 1.3.6,
     http-client-tls           >= 0.3.6,
     jose                      >= 0.10 && < 0.12,
