+++ executables/server/Aftok/Servant/Billing.hs
@@ -1,23 +1,20 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE FlexibleContexts #-}
 {-# LANGUAGE OverloadedStrings #-}
 {-# LANGUAGE TypeOperators #-}
 
 module Aftok.Servant.Billing
-  ( -- * API Types
+  ( -- * API Types (re-exported from aftok-api)
     BillingAPI,
     ProtectedBillingAPI,
     ProjectBillablesAPI,
+    BillableCreateRequest (..),
+    SubscribeRequest (..),
+    PaymentRequestCreateRequest (..),
 
     -- * Handlers
     protectedBillingServer,
     projectBillablesServer,
-
-    -- * Request/Response Types
-    BillableCreateRequest (..),
-    SubscribeRequest (..),
-    PaymentRequestCreateRequest (..),
 
     -- * JSON helpers
     billableJSON,
@@ -25,11 +22,18 @@ module Aftok.Servant.Billing
   )
 where
 
+import Aftok.API.Billing
+  ( BillableCreateRequest (..),
+    BillingAPI,
+    PaymentRequestCreateRequest (..),
+    ProjectBillablesAPI,
+    ProtectedBillingAPI,
+    SubscribeRequest (..),
+  )
 import Aftok.Billing
   ( Billable,
     Billable' (..),
     BillableId (..),
-    Recurrence (..),
     SubscriptionId,
   )
 import qualified Aftok.Billing as B
@@ -65,103 +69,23 @@ import Aftok.Payments.Types
     _PaymentRequestId,
   )
 import qualified Aftok.Payments.Zcash as Zcash
-import Aftok.Servant.App (AppM, envNetworkMode, envDbPool, runDB)
+import Aftok.Servant.App (AppM, envDbPool, envNetworkMode, runDB)
 import Aftok.Servant.Auth (AuthenticatedUser (..))
 import Aftok.Servant.Json (zip321PaymentRequestJSON)
 import Aftok.Types (ProjectId, UserId)
 import Control.Lens (to, (.~), (^.))
 import Data.Aeson
-  ( FromJSON (..),
-    Object,
-    Value (..),
-    (.:),
-    (.:?),
+  ( Value (..),
     (.=),
   )
 import qualified Data.Aeson as A
-import qualified Data.Aeson.KeyMap as O
-import Data.Aeson.Types (Pair, Parser)
+import Data.Aeson.Types (Pair)
 import Data.AffineSpace ((.+^))
 import Data.Pool (withResource)
 import qualified Data.Thyme.Clock as C
 import Data.Thyme.Time.Core (toThyme)
 import Servant
 import Servant.Auth.Server (AuthResult (..))
-
---------------------------------------------------------------------------------
--- Data Types (must be defined before API types that reference them)
---------------------------------------------------------------------------------
-
--- | Billable creation request
-data BillableCreateRequest = BillableCreateRequest
-  { bcrName :: Text,
-    bcrDescription :: Text,
-    bcrMessage :: Text,
-    bcrRecurrence :: Recurrence,
-    bcrCurrency :: Text,
-    bcrAmount :: Int,
-    bcrGracePeriod :: Int,
-    bcrRequestExpiryPeriod :: Int,
-    bcrPaymentRequestEmailTemplate :: Maybe Text,
-    bcrPaymentRequestMemoTemplate :: Maybe Text
-  }
-
-instance FromJSON BillableCreateRequest where
-  parseJSON = A.withObject "BillableCreateRequest" $ \outer -> do
-    v <- outer .: "schemaVersion"
-    when ((v :: Text) /= "1.0") $ fail "Unsupported schema version"
-    o <- outer .: "Billable"
-    BillableCreateRequest
-      <$> o .: "name"
-      <*> o .: "description"
-      <*> o .: "message"
-      <*> (parseRecurrence' =<< o .: "recurrence")
-      <*> o .: "currency"
-      <*> o .: "amount"
-      <*> o .: "gracePeriod"
-      <*> o .: "requestExpiryPeriod"
-      <*> o .:? "paymentRequestEmailTemplate"
-      <*> o .:? "paymentRequestMemoTemplate"
-
--- | Subscribe request (currently empty, billableId comes from URL)
-data SubscribeRequest = SubscribeRequest
-  deriving (Generic)
-
-instance FromJSON SubscribeRequest where
-  parseJSON _ = pure SubscribeRequest
-
--- | Payment request creation request (currently empty)
-data PaymentRequestCreateRequest = PaymentRequestCreateRequest
-  deriving (Generic)
-
-instance FromJSON PaymentRequestCreateRequest where
-  parseJSON _ = pure PaymentRequestCreateRequest
-
---------------------------------------------------------------------------------
--- API Types
---------------------------------------------------------------------------------
-
--- | Public Billing API (none currently)
-type BillingAPI = EmptyAPI
-
--- | Protected billing API for subscriptions
-type ProtectedBillingAPI =
-  "subscribe"
-    :> Capture "billableId" BillableId
-    :> ReqBody '[JSON] SubscribeRequest
-    :> Post '[JSON] SubscriptionId
-
--- | Project-specific billables API (nested under projects)
-type ProjectBillablesAPI =
-  -- GET /projects/:projectId/billables
-  Get '[JSON] Value
-    -- POST /projects/:projectId/billables
-    :<|> ReqBody '[JSON] BillableCreateRequest :> Post '[JSON] BillableId
-    -- POST /projects/:projectId/billables/:billableId/paymentRequests
-    :<|> Capture "billableId" BillableId
-      :> "paymentRequests"
-      :> ReqBody '[JSON] PaymentRequestCreateRequest
-      :> Post '[JSON] Value
 
 --------------------------------------------------------------------------------
 -- Handlers
@@ -215,19 +139,20 @@ billableCreateHandler (Authenticated user) pid req = do
     "ZEC" -> pure $ Amount ZEC (Zatoshi $ fromIntegral $ bcrAmount req)
     "BTC" -> pure $ Amount BTC (Satoshi $ fromIntegral $ bcrAmount req)
     c -> throwError err400 {errBody = "Currency " <> encodeUtf8 c <> " not recognized."}
-  let b = Billable
-        { _project = pid,
-          _creator = uid,
-          _name = bcrName req,
-          _description = Just $ bcrDescription req,
-          _messageText = Just $ bcrMessage req,
-          _recurrence = bcrRecurrence req,
-          _amount = amount,
-          _gracePeriod = bcrGracePeriod req,
-          _requestExpiryPeriod = toThyme $ fromIntegral $ bcrRequestExpiryPeriod req,
-          _paymentRequestEmailTemplate = bcrPaymentRequestEmailTemplate req,
-          _paymentRequestMemoTemplate = bcrPaymentRequestMemoTemplate req
-        }
+  let b =
+        Billable
+          { _project = pid,
+            _creator = uid,
+            _name = bcrName req,
+            _description = Just $ bcrDescription req,
+            _messageText = Just $ bcrMessage req,
+            _recurrence = bcrRecurrence req,
+            _amount = amount,
+            _gracePeriod = bcrGracePeriod req,
+            _requestExpiryPeriod = toThyme $ fromIntegral $ bcrRequestExpiryPeriod req,
+            _paymentRequestEmailTemplate = bcrPaymentRequestEmailTemplate req,
+            _paymentRequestMemoTemplate = bcrPaymentRequestMemoTemplate req
+          }
   runDB $ createBillable uid b
 billableCreateHandler _ _ _ =
   throwError err401 {errBody = "Authentication required"}
@@ -257,9 +182,11 @@ createPaymentRequestHandler cfg (Authenticated user) pid bid _ = do
           -- Run runExceptT to get QDBM (Either PaymentRequestError a)
           -- Run runQDBM to get ExceptT DBError IO (Either PaymentRequestError a)
           -- Run runExceptT to get IO (Either DBError (Either PaymentRequestError a))
-          res <- liftIO $ withResource pool $ \conn ->
-            runExceptT $ runQDBM nmode conn $ runExceptT $
-              createPaymentRequest ops now bid (b & B.amount .~ v) billDay
+          res <- liftIO $
+            withResource pool $ \conn ->
+              runExceptT $ runQDBM nmode conn $
+                runExceptT $
+                  createPaymentRequest ops now bid (b & B.amount .~ v) billDay
           case res of
             Left dbErr ->
               throwError err500 {errBody = "Database error: " <> show dbErr}
@@ -296,12 +223,12 @@ billableJSON (bid, b) =
       ]
 
 -- | Serialize recurrence to JSON
-recurrenceJSON :: Recurrence -> Value
+recurrenceJSON :: B.Recurrence -> Value
 recurrenceJSON = \case
-  Annually -> A.object ["annually" .= A.Null]
-  Monthly d -> A.object ["monthly" .= d]
-  Weekly d -> A.object ["weekly" .= d]
-  OneTime -> A.object ["onetime" .= A.Null]
+  B.Annually -> A.object ["annually" .= A.Null]
+  B.Monthly d -> A.object ["monthly" .= d]
+  B.Weekly d -> A.object ["weekly" .= d]
+  B.OneTime -> A.object ["onetime" .= A.Null]
 
 -- | Serialize amount to JSON
 amountJSON :: Amount -> Value
@@ -309,7 +236,7 @@ amountJSON (Amount ZEC (Zatoshi z)) = A.object ["currency" .= ("ZEC" :: Text), "
 amountJSON (Amount BTC (Satoshi s)) = A.object ["currency" .= ("BTC" :: Text), "satoshi" .= s]
 
 -- | Serialize payment request detail to JSON
-paymentRequestDetailJSON :: (PaymentRequestId, SomePaymentRequestDetail) -> Object
+paymentRequestDetailJSON :: (PaymentRequestId, SomePaymentRequestDetail) -> A.Object
 paymentRequestDetailJSON (rid, (SomePaymentRequest req)) =
   obj $
     ["payment_request_id" .= (rid ^. _PaymentRequestId)] <> fields req
@@ -337,29 +264,3 @@ bip70PaymentRequestJSON r =
             "payment_request_protobuf_64" .= (r ^. Bitcoin.bip70Request . to protoBase64)
           ]
     ]
-
---------------------------------------------------------------------------------
--- Parsing Helpers
---------------------------------------------------------------------------------
-
--- | Parse a recurrence value from JSON
-parseRecurrence :: Object -> Parser Recurrence
-parseRecurrence o =
-  let parseAnnually o' = const (pure Annually) <$> O.lookup "annually" o'
-      parseMonthly o' = fmap Monthly . A.parseJSON <$> O.lookup "monthly" o'
-      parseWeekly o' = fmap Weekly . A.parseJSON <$> O.lookup "weekly" o'
-      parseOneTime o' = const (pure OneTime) <$> O.lookup "onetime" o'
-      notFound =
-        fail $ "Value " <> show o <> " does not represent a Recurrence value."
-      parseV val =
-        parseAnnually val
-          <|> parseMonthly val
-          <|> parseWeekly val
-          <|> parseOneTime val
-   in fromMaybe notFound $ parseV o
-
--- | Parse a recurrence from a JSON value
-parseRecurrence' :: Value -> Parser Recurrence
-parseRecurrence' = \case
-  (Object o) -> parseRecurrence o
-  val -> fail $ "Value " <> show val <> " is not a JSON object."
