+++ executables/server/Aftok/Servant/Config.hs
@@ -1,12 +1,9 @@
 {-# LANGUAGE DataKinds #-}
-{-# LANGUAGE DeriveGeneric #-}
 {-# LANGUAGE TypeOperators #-}
 
 module Aftok.Servant.Config
-  ( -- * API
+  ( -- * API Types (re-exported from aftok-api)
     ConfigAPI,
-
-    -- * Response Types
     ClientConfig (..),
 
     -- * Server
@@ -14,22 +11,10 @@ module Aftok.Servant.Config
   )
 where
 
+import Aftok.API.Config (ClientConfig (..), ConfigAPI)
 import Aftok.ServerConfig (CaptchaConfig, captchaSiteKey)
 import Control.Lens ((^.))
-import Data.Aeson (ToJSON)
-import Servant
-
--- | Client configuration returned by the server
-data ClientConfig = ClientConfig
-  { recaptchaSiteKey :: Text
-  }
-  deriving (Show, Generic)
-
-instance ToJSON ClientConfig
-
--- | Configuration API - public endpoint for client config
-type ConfigAPI =
-  "config" :> Get '[JSON] ClientConfig
+import Servant (ServerT)
 
 -- | Config server implementation
 configServer :: Applicative m => CaptchaConfig -> ServerT ConfigAPI m
